aws: Config use pointers instead of values [Breaking Change]
Switches all primitive values to pointers to ensure configurations can be merged
correctly between service instantiation without oddities such as not being able
to set a value to its Zero because the Zero value is used to determine if the
incoming configuration parameter is set.

To facilitate this change and to clarify what aws.String function does it was
renamed to aws.StringPtr. This applied to the other pointer helper functions
as well. An IntPtr was also added for *int.

    String  -> StringPtr
    Boolean -> BoolPtr
            -> IntPtr // New
    Long    -> Int64Ptr
    Double  -> Float64Ptr
    Time    -> TimePtr

In addition helper functions were added for each type to return the value that
a pointer points to or the type's zero value if the pointer is nil. This saves
some of the hassle of checking if a field is nil before dereferencing it in
circumstances where nil should be treated as zero.

To update to this change the above table of aws.String -> aws.StringPtr function
usages need to be updated to their new name aws.String being the most common.
Then any usages of aws.Config{} in your code should be updated to wrap the set
values with the relevant pointer function.

Example:

     // Code prior to update:
     svc := s3.New(aws.Config{
          Region:           "us-west-2",
          MaxRetries:       2,
          S3ForcePathStyle: true,
     }
     // Prior code read Config.Region value
     if svc.Config.Region == "us-west-2" {...}

     // After Updating to this change:
     svc := s3.New(aws.Config{
          Region:           aws.StringPtr("us-west-2"),
          MaxRetries:       aws.IntPtr(2),
          S3ForcePathStyle: aws.BoolPtr(true),
     }
     // After this change reading Config.Region value
     if aws.StringPtrValue(svc.Config.Region) == "us-west-2" {...}

Fixes #157